requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-calendar>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
easyocr>=1.7.0
Pillow>=10.0.0
openai>=1.0.0
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0
python-dotenv>=1.0.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.88.0
requests>=2.31.0
openpyxl>=3.1.0